requirements.txt
Django
Pillow
django-filter
djangorestframework
djangorestframework-simplejwt
django-jazzmin
drf-spectacular
drf-spectacular-sidecar
django-rest-swagger
django-cors-headers
django-celery-beat